automation/requirements.txt
langchain
langchain-openai
chromadb
typer
sentence-transformers
pypdf
python-dotenv
rich
